+++ custom_components/wundergroundpws/__init__.py
@@ -1,5 +1,5 @@
-'''The wundergroundpws component.'''
-
+"""The wundergroundpws component."""
+import json
 import logging
 
 import voluptuous as vol
@@ -17,15 +17,16 @@ from .const import (
     CONF_LANG,
     CONF_NUMERIC_PRECISION,
     CONF_PWS_ID,
+    CONF_CALENDARDAYTEMPERATURE,
 
     DOMAIN,
-    NAME,
 
     ENTRY_PWS_ID,
     ENTRY_WEATHER_COORDINATOR,
 
     LANG_CODES,
-    DEFAULT_LANG,
+    DEFAULT_LANG, ENTRY_LANG, ENTRY_TRAN_FILE,
+    DEFAULT_CALENDARDAYTEMPERATURE, ENTRY_CALENDARDAYTEMPERATURE, DEFAULT_NUMERIC_PRECISION
 )
 from .wunderground_data import WUndergroundData
 
@@ -36,11 +37,12 @@ CONFIG_SCHEMA = vol.Schema(
         DOMAIN: {
             vol.Required(CONF_API_KEY): cv.string,
             vol.Required(CONF_PWS_ID): cv.string,
-            vol.Required(CONF_NUMERIC_PRECISION):
+            vol.Optional(CONF_NUMERIC_PRECISION, default=DEFAULT_NUMERIC_PRECISION):
                 vol.All(vol.In(['none', 'decimal'])),
-
             vol.Optional(CONF_LANG, default=DEFAULT_LANG):
                 vol.All(vol.In(LANG_CODES)),
+            vol.Optional(CONF_CALENDARDAYTEMPERATURE, default=DEFAULT_CALENDARDAYTEMPERATURE):
+                cv.boolean,
             vol.Inclusive(CONF_LATITUDE, 'coordinates',
                           'Latitude and longitude must exist together'):
                 cv.latitude,
@@ -53,8 +55,22 @@ CONFIG_SCHEMA = vol.Schema(
 )
 
 
+"""get translation file for wupws sensor friendly_name"""
+def get_tran_file(hass: HomeAssistantType):
+    tfiledir = f'{hass.config.config_dir}/custom_components/{DOMAIN}/wupws_translations/'
+    tfilename = hass.data[DOMAIN][ENTRY_LANG].split('-', 1)[0]
+    try:
+        tfile = open(f'{tfiledir}{tfilename}.json')
+        tfiledata = json.load(tfile)
+    except:
+        tfile = open(f'{tfiledir}en.json')
+        tfiledata = json.load(tfile)
+        _LOGGER.warning(f'Sensor translation file {tfilename}.json does not exist. Defaulting to en-US.')
+    return tfiledata
+
+
 async def async_setup(hass: HomeAssistantType, config: ConfigType) -> bool:
-    '''Set up the WUnderground sensor.'''
+    """Set up the WUnderground sensor."""
 
     platform_config = config.get(DOMAIN)
 
@@ -67,6 +83,7 @@ async def async_setup(hass: HomeAssistantType, config: ConfigType) -> bool:
     numeric_precision = platform_config.get(CONF_NUMERIC_PRECISION)
 
     lang = platform_config.get(CONF_LANG)
+    caldaytemp = platform_config.get(CONF_CALENDARDAYTEMPERATURE)
 
     if hass.config.units is METRIC_SYSTEM:
         unit_system_api = 'm'
@@ -89,6 +106,10 @@ async def async_setup(hass: HomeAssistantType, config: ConfigType) -> bool:
     hass.data[DOMAIN] = {
         ENTRY_WEATHER_COORDINATOR: rest,
         ENTRY_PWS_ID: pws_id,
+        ENTRY_LANG: lang,
+        ENTRY_CALENDARDAYTEMPERATURE: caldaytemp
     }
+
+    hass.data[DOMAIN].update({ENTRY_TRAN_FILE: get_tran_file(hass)})
 
     return True